+++ src/libsystemd/rust/sd_event.rs
@@ -7,9 +7,9 @@ use crate::event_source_type::EventSourceType;
 
 pub type Result<T> = std::result::Result<T, i32>;
 
-pub const NEG_EBADF: i32 = -(libc::EBADF as i32);
-pub const NEG_EINVAL: i32 = -(libc::EINVAL as i32);
-pub const NEG_ENODATA: i32 = -(libc::ENODATA as i32);
+pub const NEG_EBADF: i32 = -libc::EBADF;
+pub const NEG_EINVAL: i32 = -libc::EINVAL;
+pub const NEG_ENODATA: i32 = -libc::ENODATA;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum EventState {
